requirements.txt
# 데이터 수집
pykrx>=1.0.45
finance-datareader>=0.9.50
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=10.0.0
scipy>=1.11.0

# 한국투자증권 API
requests>=2.31.0

# 스케줄러
APScheduler>=3.10.0,<4.0.0

# 유틸
python-dotenv>=1.0.0
loguru>=0.7.0
